sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "evalyn-sdk"
version = "0.0.1"
description = "Auto-evaluation toolkit for instrumenting LLM agents, collecting traces, and running metrics."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [{ name = "Evalyn", email = "dev@evalyn.local" }]
dependencies = []

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23"
]
llm = [
  "openai>=1.53.0,<2.0",
  "google-generativeai>=0.7.2,<1.0"
]
otel = [
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0"
]
agent = [
  "langchain-core>=0.3.0,<0.4",
  "langgraph>=0.1.14,<1.0",
  "langchain-google-genai>=2.0.5,<3.0",
  "google-genai>=0.2.0,<1.0",
  "python-dotenv>=1.0.1,<2.0",
  "pydantic>=2.6.0,<3.0",
]
# Note: google-adk requires google-genai>=1.14.0 which conflicts with agent extra's
# google-genai<1.0. Users should install google-adk separately if needed.
# google-adk = ["google-adk>=1.0.0", "opentelemetry-sdk>=1.27.0"]
anthropic-agents = [
  "anthropic>=0.40.0",
]

[project.scripts]
evalyn = "evalyn_sdk.cli:main"

[tool.setuptools.packages.find]
include = ["evalyn_sdk*"]
